improve tests and move implementation to emio's headers
