+++ include/emio/buffer.hpp
@@ -449,6 +449,53 @@ class iterator_buffer<std::back_insert_iterator<Container>> final : public buffe
 template <typename Iterator>
 iterator_buffer(Iterator&&) -> iterator_buffer<std::decay_t<Iterator>>;
 
+/**
+ * This class fulfills the buffer API by using a file stream and an internal cache.
+ */
+class file_buffer : public buffer<char> {
+ public:
+  /**
+   * Constructs and initializes the buffer with the given file stream.
+   * @param file The file stream.
+   */
+  constexpr explicit file_buffer(std::FILE* file) : f_{file} {
+    this->set_write_area(cache_);
+  }
+
+  file_buffer(const file_buffer&) = delete;
+  file_buffer(file_buffer&&) = delete;
+  file_buffer& operator=(const file_buffer&) = delete;
+  file_buffer& operator=(file_buffer&&) = delete;
+
+  /**
+   * Flushes the internal cache to the file stream.
+   * @note Does not flush the file stream!
+   */
+  constexpr result<void> flush() noexcept {
+    const size_t written = std::fwrite(cache_.data(), sizeof(char), this->get_used_count(), f_);
+    if (written != this->get_used_count()) {
+      return err::eof;
+    }
+    this->set_write_area(cache_);
+    return success;
+  }
+
+ protected:
+  constexpr result<std::span<char>> request_write_area(const size_t /*used*/, const size_t size) noexcept override {
+    EMIO_TRYV(flush());
+    const std::span<char> area{cache_};
+    this->set_write_area(area);
+    if (size > cache_.size()) {
+      return area;
+    }
+    return area.subspan(0, size);
+  }
+
+ private:
+  std::FILE* f_;
+  std::array<char, detail::internal_buffer_size> cache_;
+};
+
 namespace detail {
 
 /**
